chore: enable `pre-post-scripts` and add `prestart` command
A new line was added to the .npmrc file to enable the execution of pre and post scripts. This was essential to add a new "prestart" command in the package.json file that runs the build command before starting the server. The change ensures that the most recent build will always be used when starting the server.