feat: add LinguiJS for i18n and update dependencies
Added LinguiJS for handling internationalization (i18n). This includes the addition of new devDependencies "@lingui/cli", "@lingui/conf", "@lingui/loader", "@lingui/macro", "@lingui/react", and "@lingui/swc-plugin". Additionally, added the "lingui:extract" and "lingui:compile" commands to the "scripts" section in package.json for handling the generation and compiling of translation files.

The "pnpm-lock.yaml" was updated to accommodate for the new dependencies and to reflect updates in the versions of the other dependencies. The updates in the dependency versions were necessary to assure compatibility with the newly added LinguiJS.