chore: add Prettier and its plugins for code formatting
This commit introduced Prettier, a code formatter, along with some of its plugins for task automation. The eslint-config-prettier is used to ensure that ESLint and Prettier do not conflict, allowing them to be used together seamlessly. The prettier-plugin-organize-imports is utilised to automatically sort and organise imports, making the code much more digestible and maintaining the consistency in the codebase. The prettier-plugin-tailwindcss helps sort and organise tailwind classes which ultimately enhances the readability of the code. The '@types/node' was also downgraded from 20.4.2 to 20.3.3 to accommodate for these additional formatting plugins.

The purpose of these changes is to improve code quality, readability, and consistency throughout the project; thereby, smooth collaboration among team members in code reviews or further development.