chore: add `.prettierignore` file to exclude unnecessary directories and files from formatting
This commit introduces a .prettierignore file to let Prettier know what directories and files it should not format. With this change in place, we don't run unnecessary formatting on directories and files such as the node_modules, the build folder, local environment files, and log files among others. This will help us keep an optimized workflow while maintaining codebase consistency.