chore(deps): add `@lingui/detect-locale` package
Added @lingui/detect-locale@4.2.1 to the dependency list in the package.json file and in the pnpm-lock.yaml file. This new module was added to provide a feature to automatically detect the user's locale.

Downgraded the @types/node version from 20.4.5 to 20.3.3 in the pnpm-lock.yaml file due to compatibility issues in the current environment.