ci: add initial commit check to build workflow
A new job named 'check' was added to the GitHub Actions build workflow that checks if the current commit is the initial commit of the repository. If it is, the 'Build & Test' job will be skipped. This is done to avoid unnecessary runs for the initial commit as typically it does not contain substantial code to build or test.