ci: add state verification to GitHub workflow
This commit introduces an enhancement into the .github/workflows/build.yml. The changes made include the addition of a new job entitled 'State Verifier'. Its purpose is to verify the state of the repository code by checking whether the commit is the initial commit. This is valuable for determining if continuous integration should be run or not. A new git command was added to count the number of revisions, and an environment variable "skip_ci" was introduced and is set based on the outcome of the count.