chore: update line references in localization files
This commit updates the line references for translations in the Arabic, English, Farsi, and Turkish localization files. These were changed due to modifications in the structure of the related components. Any time the components are modified, the line references in all localization files need to be updated to ensure that the translations are correctly linked.