chore: update locales reference lines in `messages.po` files
The lines indicating the location of each msgid in the project were outdated. They have been corrected to point to the right lines in the right files. The changes involve no modification in the translations themselves, just the row reference in the codebase. Each messages.po file in every 'locales' directory: 'ar', 'en', 'fa', 'tr' was updated.