refactor: update global font styling in Layout component
In src/components/layout.tsx the styling was previously applied to the <main> tag only. To improve consistency and ensure the chosen font is applied across the entire application, the styling has been moved to html in a global styling block. This commit will promote uniformity in font style across the website.