chore(deps): update dependencies with new react-use hooks and types
This commit contains a significant change in dependencies of the application. It removes `usehooks-ts`, which was having version 2.9.1 and replaces it with `react-use`, having version 17.4.0. This is done to use the newer hooks provided by react-use that offer more functionalities and are optimized for performance. Furthermore, several new type definitions have been added like `@types/js-cookie@2.2.7` and `/@xobotyi/scrollbar-width@1.9.5` to improve type checking, code reliability, and intellisense in IDE.