refactor: replace `useIsMounted` with `useMountedState` from react-use
The useIsMounted hook has been replaced by useMountedState from the 'react-use' library. This change was carried out to upgrade to a more performant and recommended way to check if a component is still mounted before setting state, which can prevent memory leaks. Additionally, this can lead to less ambiguity and increase efficiency in code execution.