chore(deps): update package dependencies and address redundancy
Updated versions of various packages in pnpm-lock.yaml to foster compatibility and efficiency. Downgraded 'jiti' from 1.20.0 to 1.18.2 and '@types/eslint' from 8.44.8 to 8.44.6 to align with the latest type definitions. '@types/node' package declaration was removed from one of the dependencies as it was redundant. The 'undici-types' package was also downgraded from version 5.26.5 to 5.26.4 for compatibility.